VMC/FlightSoftware/apriltag/python/requirements.txt
colored==1.4.2
loguru==0.5.3
numpy==1.19.5
paho-mqtt==1.5.1
transforms3d==0.3.1